+++ src/main/kotlin/com/sparta/bubbleclub/domain/keyword/entity/KeywordStore.kt
@@ -1,12 +1,15 @@
 package com.sparta.bubbleclub.domain.keyword.entity
 
 import jakarta.persistence.*
+import org.hibernate.annotations.CreationTimestamp
+import java.time.ZonedDateTime
 
 @Entity
 @Table(name = "keywords")
 class KeywordStore(
     keyword: String,
-    count: Long
+    count: Long,
+    createdAt: ZonedDateTime
 ) {
 
     @Id
@@ -19,6 +22,10 @@ class KeywordStore(
 
     @Column(name = "count")
     var count: Long = count
+
+    @Column(name = "created_at", nullable = false, updatable = false, columnDefinition = "TIMESTAMP(6)")
+    @CreationTimestamp
+    var createdAt: ZonedDateTime = createdAt
 
     fun increaseCount() = this.count++
 }
