+++ src/main/kotlin/com/sparta/bubbleclub/global/cache/CachingConfig.kt
@@ -1,4 +1,4 @@
-package com.sparta.bubbleclub.global.config
+package com.sparta.bubbleclub.global.cache
 
 import org.springframework.cache.annotation.EnableCaching
 import org.springframework.context.annotation.Configuration
