+++ src/main/kotlin/com/sparta/bubbleclub/global/entity/JpaAuditingConfig.kt
@@ -1,4 +1,4 @@
-package com.sparta.bubbleclub.global.config
+package com.sparta.bubbleclub.global.entity
 
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing
