+++ src/main/kotlin/com/sparta/bubbleclub/global/schedule/SchedulingConfig.kt
@@ -1,0 +1,8 @@
+package com.sparta.bubbleclub.global.schedule
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.scheduling.annotation.EnableScheduling
+
+@EnableScheduling
+@Configuration
+class SchedulingConfig
