+++ src/main/kotlin/com/sparta/bubbleclub/infra/jwt/JwtProperties.kt
@@ -1,4 +1,4 @@
-package com.sparta.bubbleclub.global.security.jwt
+package com.sparta.bubbleclub.infra.jwt
 
 import org.springframework.boot.context.properties.ConfigurationProperties
